feat: further modularilze sdk
continue splitting SDK into smaller manageable chunks

- split auth into authtoken and set of individual modules supposed to
  handle related auth methods
- refactor AuthError (split out AuthTokenError) and normalize error
  names
