+++ openstack_sdk/README.md
@@ -15,8 +15,8 @@ additional convenience features.
 - `Query`, `Find` and `Pagination` interfaces implementing basic functionality
 - `RawQuery` interface providing more control over the API
   invokation with upload and download capabilities.
-- Every combination of URL + http method + body schema is
-- represented by a dedicated module
+- Every combination of URL + http method + body schema is represented by a
+dedicated module
 
 ## Structure
 
@@ -61,3 +61,9 @@ async fn list_flavors() -> Result<(), OpenStackError> {
     Ok(())
 }
 ```
+
+## Documentation
+
+Crate documentation [is published here](https://docs.rs/openstack_sdk)
+
+Project documentation [is available here](https://gtema.github.io/openstack)
